refactor(cache): split god module by concern
Carve the 3417-line cache.ts into seven focused modules under cache/,
layered as a dependency DAG with no cycles:

  db.ts          cache paths, shared query helpers, FTS integrity guard
  messages.ts    row ↔ domain mapping, prepared statements, text builders
  schema.ts      DDL, migrations, withCacheDb/ensureSchema (deep interface)
  sessions.ts    load / save / clear / info / initialization
  search.ts      FTS index sync, query, snippet/highlight
  file-activity.ts  per/cross-session file activity + path search
  project-groups.ts project grouping view query

cache.ts becomes a thin barrel re-exporting the same 17 symbols + 9 types,
so discovery/index.ts and all callers are unchanged. Shared row types
(CacheRow, ScalarRow, SessionHeadChange) were lifted to db.ts to keep the
layering acyclic. Pure textual relocation — no function body, signature,
SQL, or on-disk format changes.
